split sensor_ic2.h
