+++ release/files_common
@@ -318,8 +318,9 @@ system/camerad/main.cc
 system/camerad/snapshot/*
 system/camerad/cameras/camera_common.h
 system/camerad/cameras/camera_common.cc
-system/camerad/cameras/sensor2_i2c.h
 system/camerad/sensors/ar0231.cc
+system/camerad/sensors/ar0231_registers.h
+system/camerad/sensors/ox03c10_registers.h
 
 selfdrive/manager/__init__.py
 selfdrive/manager/build.py
